refactor: keep components specific to a task
